requirements.txt
streamlit==1.8.0
pandas==1.3.3
pyzbar==0.1.8
pillow==8.3.1
openpyxl==3.0.7